% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_set}
\alias{dummy_set}
\title{Tech Stock Time Series Dataset}
\format{
A data frame with 2133 observations of 4 variables:
\describe{
  \item{dates}{Character vector of dates in "YYYY-MM-DD" format.}
  \item{TSLA.Close}{Numeric. Closing prices for Tesla.}
  \item{MSFT.Close}{Numeric. Closing prices for Microsoft.}
  \item{MARA.Close}{Numeric. Closing prices for MARA Holdings.}
}
}
\usage{
data(dummy_set)
}
\description{
A multivariate dataset for closing prices for several major tech stocks over time. Source: YahooFinance.
}
\examples{
data(dummy_set)
plot(as.Date(dummy_set$dates), dummy_set$TSLA.Close, type = "l")
}
\keyword{datasets}
